Move `FontSettings`, `FontStack` to `styled_text`
