+++ styled_text/src/lib.rs
@@ -23,8 +23,12 @@ extern crate alloc;
 
 mod attributes;
 mod font_family;
+mod font_settings;
+mod font_stack;
 mod generic;
 
 pub use attributes::{Stretch, Style, Weight};
 pub use font_family::FontFamily;
+pub use font_settings::FontSettings;
+pub use font_stack::FontStack;
 pub use generic::GenericFamily;
